Move function to CustomerDataAccess
